TREZOR based recovery mode.
This is an experimental patch that allows recovering from a pass phrase
without revealing any information to the computer.  The seed words are
chosen on the TREZOR using the buttons as prev/next.  The computer just
sends an empty string to indicate that the right word was chosen.

Currently recovery only works via python-trezor.  MyTrezor will timeout
and abort recovery.  Use python-trezor as:

./cmdtr.py recovery_device -w 24 -p -r -l label